backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
httpx
python-jose[cryptography]
passlib[bcrypt]
email-validator
